+++ L_AutoVirtualThermostat1.lua
@@ -1428,7 +1428,7 @@ function plugin_init(dev)
 
     -- Check for ALTUI and OpenLuup
     for k,v in pairs(luup.devices) do
-        if v.device_type == "urn:schemas-upnp-org:device:altui:1" then
+        if v.device_type == "urn:schemas-upnp-org:device:altui:1" and v.device_num_parent == 0 then
             local rc,rs,jj,ra
             D("init() detected ALTUI at %1", k)
             isALTUI = true