+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 1.6 (stable branch) ##
 
+* Fix so odd return values from underlying Luup requests don't crash Status Data request.
 * Experimental fix for openLuup switch control (stable-180918)
 * openLuup loads modules differently from Vera Luup, so work out how to get AVT working on openLuup when multiple device instances are created (differences don't matter if only one device).