+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 1.6 (develop branch) ##
 
+* Upgrade detection of AltUI so we don't falsely detect when bridged (on "real" device triggers AltUI feature registration).
 * Implement AllSetpoints state variable, for user convenience. 
 * openLuup loads modules differently from Vera Luup, so work out how to get AVT working on openLuup when multiple device instances are created (differences don't matter if only one device).