+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change Log #
 
-## Version 1.6 (stable branch) ##
+## Version 1.6 (development)
 
 * Fix so odd return values from underlying Luup requests don't crash Status Data request.
 * Experimental fix for openLuup switch control (stable-180918)
+* Upgrade detection of AltUI so we don't falsely detect when bridged (on "real" device triggers AltUI feature registration).
+* Implement AllSetpoints state variable, for user convenience. 
 * openLuup loads modules differently from Vera Luup, so work out how to get AVT working on openLuup when multiple device instances are created (differences don't matter if only one device).
 
 ## Version 1.5 (released) ##